Move NormaliseMappings to config package
